13 hours.
Ok, bullets work so far, enemies got hurt and die (die effect's graphics present, but not coded it yet).

+++ src/es/ld23/util/map/PC.java
@@ -84,7 +84,7 @@ public abstract class PC {
 	public void cancelMovement(long delta) {
 	}
 	
-	public boolean hurt(int dmg) {
+	public boolean hurt(double dmg) {
 		return false;
 	}
